+++ R/clusteredinterference.R
@@ -1,14 +1,14 @@
 #' Draw Causal Inference from Observational Studies with Clustered Interference
 #'
-#' \pkg{clusteredinterference} implements the estimators from Barkley et al.,
-#' (201X), \emph{Causal Inference from Observational Studies with Clustered
-#' Interference}. These are inverse probability of treatment-weighted estimators
-#' for estimating the causal effects on outcomes of population-level treatment
-#' policies.
+#' \pkg{clusteredinterference} implements the estimators from Barkley et al.
+#' (2017) \emph{Causal Inference from Observational Studies with Clustered
+#' Interference} <arXiv:1711.04834>. These are inverse probability of
+#' treatment-weighted estimators for estimating the causal effects on outcomes
+#' of population-level treatment policies.
 #'
 #' These estimators are implemented in one function, \code{\link{policyFX}}. A
 #' healthy amount of documentation can be found in the vignette included in this
-#' package.
+#' package: \code{vignette("estimate-policyFX")}.
 #'
 #' @name clusteredinterference
 #' @docType package